CyberFriend_bot_plugin/requirements.txt
SQLAlchemy
nb-cli
nonebot2[fastapi]
nonebot-adapter-console
nonebot-adapter-onebot
nonebot_plugin_apscheduler